backend/pyproject.toml
[project]
name = "testt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-rest-auth>=7.0.1",
    "django>=5.2.1",
    "django-allauth>=65.7.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "faker>=37.1.0",
    "mysqlclient>=2.2.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "whitenoise>=6.9.0",
]
